refactor: move dapp injected code to another file
